+++ debug/access-token-client.js
@@ -49,6 +49,18 @@ const {
   // users
 } = require('./modules/import-config.js');
 
+//
+// Check if OAuth 2.0 grant type client credentials grant is disabled in configuration
+//
+if (config.oauth2.disableClientGrant) {
+  // Yes, abort the test without error
+  console.log('\nTest skipped, client credentials grant disabled in configuration.');
+  console.log('---------------------');
+  console.log('  All Tests Passed');
+  console.log('---------------------');
+  process.exit(0);
+}
+
 const managedFetch = require('./modules/managed-fetch').managedFetch;
 
 const {