+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add demo for OAuth 2.0 implicit grant type
 - Add demo for OAuth 2.0 password grant type
 - Change load-test-introspect.js to generate access token by code grant instead of client credentials grant.
+- Disable client token tests when client credentials grant is disabled
 
 ## 2024-11-24